requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
reportlab==4.0.4
python-dotenv==1.0.0
Werkzeug==2.3.7
